winRBackend/requirements.txt
Django==3.2.12
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
mysqlclient==2.1.0
python-dotenv==0.20.0
django-cors-headers==3.7.0
django-storages==1.13.2
PyJWT==2.6.0
djangorestframework-simplejwt==5.2.2
pymongo==4.3.3
bcrypt==4.0.1
PyJWT==2.6.0
django-filter==23.2
bs4==0.0.1
pandas==1.5.3
loguru==0.7.0c
pya3==1.0.29
pyotp==2.9.0
pycryptodome==3.18.0
apscheduler==3.10.4